requirements.txt
discord.py==1.7.3
aiohttp==3.7.4.post0
gidgethub==5.0.1
python-dotenv==0.17.0
fuzzywuzzy==0.18.0
motor==2.5.0
dnspython==2.1.0
psutil==5.8.0
dblpy==0.4.0
dlabs.py
statcord.py
beautifulsoup4==4.9.3
colorama==0.4.4
requests==2.26.0
py-carbon
sentry-sdk==1.3.0
aiofiles==0.7.0
plotly==5.1.0
pandas==1.3.1
kaleido
certifi==2020.12.5